pyproject.toml
[tool.poetry]
name = "test-team-up"
version = "0.1.0"
description = ""
authors = ["Zhoomart <zh@mancho.dev>"]
readme = "README.md"
packages = [{ include = "test_team_up" }]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
flake8 = "^6.0.0"
django = "3.2.4"
django-debug-toolbar = "^4.1.0"
ruff = "^0.0.270"
drf-spectacular = "^0.26.2"
drf-spectacular-sidecar = "^2023.6.1"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff] # https://github.com/charliermarsh/ruff#configuration
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations"
]

line-length = 119

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"